# docs/swagger.yaml
basePath: /api/v1
definitions:
  analysis.AnalysisRequest:
    properties:
      transactions:
        items:
          $ref: '#/definitions/analysis.Transaction'
        type: array
    type: object
  analysis.AnalysisResponse:
    properties:
      by_category:
        items:
          $ref: '#/definitions/analysis.CategorySummary'
        type: array
      by_description:
        items:
          $ref: '#/definitions/analysis.DescriptionSummary'
        type: array
      net_balance:
        type: number
      total_income:
        type: number
      total_spent:
        type: number
      transaction_count:
        type: integer
    type: object
  analysis.CategorySummary:
    properties:
      average:
        type: number
      category:
        type: string
      count:
        type: integer
      total:
        type: number
    type: object
  analysis.DescriptionSummary:
    properties:
      count:
        type: integer
      description:
        type: string
      total:
        type: number
    type: object
  analysis.Transaction:
    properties:
      amount:
        type: number
      category:
        type: string
      date:
        type: string
      description:
        type: string
    type: object
  auth.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/auth.User'
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  auth.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
    type: object
  parser.Transaction:
    properties:
      amount:
        type: number
      category:
        type: string
      date:
        type: string
      description:
        type: string
    type: object
  parser.UploadResponse:
    properties:
      count:
        type: integer
      message:
        type: string
      transactions:
        items:
          $ref: '#/definitions/parser.Transaction'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: API para análise de faturas do Nubank
  title: Gastei Quanto API
  version: "1.0"
paths:
  /analysis/transactions:
    post:
      consumes:
      - application/json
      description: Agrupa e analisa transações por categoria e descrição
      parameters:
      - description: Lista de transações
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/analysis.AnalysisRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/analysis.AnalysisResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Analisa transações
      tags:
      - analysis
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      description: Logout user (client should delete token)
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/me:
    get:
      description: Get authenticated user information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register new user
      tags:
      - auth
  /parser/upload/csv:
    post:
      consumes:
      - multipart/form-data
      description: Faz upload de um arquivo CSV contendo transações
      parameters:
      - description: CSV file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/parser.UploadResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload CSV
      tags:
      - parser
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"